fixed few more test cases
